Filtered Garbage
Maybe missed some text along the way but this is way easier to look through now.

 //0x1E's PREVIOUS byte seems to always follow the format of 0x0F, where the high byte is 0.  This removes lots of bogus shit.

Also skipping the first "match" when it finds the block text character 0x1e , it skips a lot of garbage at beginning of file.

Found an issue with some files like A_624 and C_010 where this method fails to get text.